requirements.txt
allure-pytest==2.15.3
allure-python-commons==2.15.3
attrs==25.4.0
certifi==2026.1.4
charset-normalizer==3.4.4
execnet==2.1.2
h11==0.16.0
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
MarkupSafe==3.0.3
outcome==1.3.0.post0
packaging==26.0
pluggy==1.6.0
Pygments==2.19.2
PySocks==1.7.1
pytest==9.0.2
pytest-html==4.2.0
pytest-metadata==3.1.1
pytest-rerunfailures==16.1
pytest-xdist==3.6.1
python-dotenv==1.2.1
requests==2.32.5
selenium==4.41.0
-e git+https://github.com/je80sand/selenium_automation_project.git@4a88944eccf81b97aa452414628b3e7f8353e7ec#egg=selenium_automation_project
setuptools==82.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.33.0
trio-websocket==0.12.2
typing_extensions==4.15.0
urllib3==2.6.3
webdriver-manager==4.0.2
websocket-client==1.9.0
wheel==0.46.3
wsproto==1.3.2
